+++ types/Constants.d.ts
@@ -16,37 +16,3 @@ export enum SearchTypes {
     youtube = 'ytsearch',
     youtubemusic = 'ytmsearch'
 }
-
-export type LavalinkSource = "youtube" | "youtubemusic" | "soundcloud";
-export type Snowflake = `${bigint}` | string;
-export type Base64String = string;
-
-export interface ShoukakuOptions {
-    resumable?: boolean;
-    resumableTimeout?: number;
-    reconnectTries?: number;
-    moveOnDisconnect?: boolean;
-    restTimeout?: number;
-    reconnectInterval?: number;
-    userAgent?: string;
-}
-
-export interface NodeOptions {
-    name: string;
-    url: string;
-    auth: string;
-    secure?: boolean;
-    group?: string;
-}
-
-export interface DecodedTrack {
-    identifier?: string;
-    isSeekable?: boolean;
-    author?: string;
-    length?: number;
-    isStream?: boolean;
-    position?: number;
-    title?: string;
-    uri?: string;
-    sourceName?: string;
-}
