+++ types/guild/ShoukakuConnection.d.ts
@@ -1,14 +1,11 @@
 import { EventEmitter } from 'events';
-import { state, Snowflake } from '../Constants';
-import { ShoukakuPlayer } from './ShoukakuPlayer';
+import { ConnectOptions, Snowflake, VoiceServerUpdate, VoiceStateUpdate } from '..';
+import { state } from '../Constants';
 import { ShoukakuSocket } from '../node/ShoukakuSocket';
+import { ShoukakuPlayer } from './ShoukakuPlayer';
 
 export class ShoukakuConnection extends EventEmitter {
-  constructor(
-    player: ShoukakuPlayer,
-    node: ShoukakuSocket,
-    options: { guildId: Snowflake, shardId: number }
-  );
+  constructor(player: ShoukakuPlayer, node: ShoukakuSocket, options: { guildId: Snowflake, shardId: number });
 
   public player: ShoukakuPlayer;
   public node: ShoukakuSocket;
@@ -22,8 +19,6 @@ export class ShoukakuConnection extends EventEmitter {
   public state: state.CONNECTED | state.CONNECTING | state.DISCONNECTED | state.DISCONNECTING;
   public moved: boolean;
   public reconnecting: boolean;
-  private serverUpdate: boolean;
-
   public setDeaf(deaf?: boolean): void;
   public setMute(mute?: boolean): void;
   public disconnect(): void;
@@ -32,25 +27,5 @@ export class ShoukakuConnection extends EventEmitter {
   protected setStateUpdate(options: VoiceStateUpdate): void;
   protected setServerUpdate(data: VoiceServerUpdate): void;
   protected send(d: object, important?: boolean): void;
-}
-
-export interface ConnectOptions {
-  guildId: Snowflake;
-  shardId: number;
-  channelId: Snowflake;
-  deaf?: boolean;
-  mute?: boolean;
-}
-
-export interface VoiceServerUpdate {
-  token: string;
-  guild_id: Snowflake;
-  endpoint: string | null;
-}
-
-export interface VoiceStateUpdate {
-  session_id: string;
-  channel_id: Snowflake;
-  self_deaf: boolean;
-  self_mute: boolean;
+  private serverUpdate: boolean;
 }
