--- types/libraries/Constants.d.ts
@@ -1,7 +1,0 @@
-import { Snowflake } from '../Constants';
-
-export interface GetterObj {
-    guilds: Map<any, any>;
-    id: () => Snowflake;
-    ws: (shardId: number, payload: string, important: boolean) => any;
-}
