+++ types/libraries/wrappers/Eris.d.ts
@@ -1,6 +1,5 @@
+import { GetterObj, NodeOptions } from '../..';
 import { Shoukaku } from '../../Shoukaku';
-import { NodeOptions } from '../../Constants';
-import { GetterObj } from '../Constants';
 
 export class Eris {
     constructor(
