+++ types/node/ShoukakuQueue.d.ts
@@ -1,13 +1,10 @@
 import { ShoukakuSocket } from './ShoukakuSocket';
 
 export class ShoukakuQueue {
-    constructor(
-        socket: ShoukakuSocket
-    );
+    constructor(socket: ShoukakuSocket);
 
     public socket: ShoukakuSocket;
     public pending: string[];
-
     public enqueue(data: Object, important?: boolean): void;
     public process(): void;
     public clear(): void;
